% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters_ui}
\alias{filters_ui}
\title{Filters, dropdowns ui}
\usage{
filters_ui(id, .salaries, ...)
}
\description{
Filters, dropdowns ui
}
\examples{
\dontrun {

  ui <- page_sidebar(theme = bs_theme(
      version = 5,  fg = '#EEE8D5', bg =  '#232627', primary = 'seagreen'),
    #actionButton('poo', 'poo', class = 'btn-secondary'),
    #selectInput('select', 'select', choices = '1')
    filters_ui('poo', salaries)
  )
  
 shinyApp(ui = ui, server = function(input, output) {})

}
}
