% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_dark_mode.R
\name{gt_dark_mode}
\alias{gt_dark_mode}
\title{stylize to dark mode}
\usage{
gt_dark_mode(gt_object, ...)
}
\arguments{
\item{gt_object}{gt table}

\item{...}{tab options}
}
\value{
gt table
}
\description{
stylize to dark mode
}
