% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{gt_salary_stats}
\alias{gt_salary_stats}
\title{Create stats table for salary data}
\usage{
gt_salary_stats(.df, .return_data = FALSE, .gt = TRUE)
}
\arguments{
\item{.df}{salaries}
}
\description{
Create stats table for salary data
}
\examples{
data(salaries)
salaries |> filter(title_general == 'Scientist') |>   gt_salary_stats()

}
