% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R, R/server.R, R/ui.R
\name{launch_app}
\alias{launch_app}
\alias{rb_server}
\alias{rb_ui}
\title{launch rbiotechsalary app}
\usage{
launch_app(options = list(), ui = rb_ui, server = rb_server, ...)

rb_server(input, output, session)

rb_ui()
}
\description{
launch rbiotechsalary app
}
\section{Functions}{
\itemize{
\item \code{rb_server()}: updated server function

\item \code{rb_ui()}: UI components

}}
\examples{
\dontrun{

 launch_app()
  
 # launch ui without server
 launch_app(server = function(input, output) {})
 
}
\dontrun{

  # launch with just ui
  launch_app(ui = rb_ui, server = function(input, output) {})

  # launch with server
  launch_app(ui = rb_ui, server = rb_server)
}
}
