% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grouped_options.R
\name{make_grouped_options_2}
\alias{make_grouped_options_2}
\title{makes grouped options list for selectize input}
\usage{
make_grouped_options_2(.df, .groups, .choices)
}
\description{
makes grouped options list for selectize input
}
\examples{
make_grouped_options_2(salaries, title_category, title_general)
}
