% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_basic_stats.R
\name{plot_basic_salary_stats}
\alias{plot_basic_salary_stats}
\title{small plot salary stats}
\usage{
plot_basic_salary_stats(.df)
}
\arguments{
\item{.df}{salaries}
}
\description{
small plot salary stats
}
\examples{
data(salaries)
salaries |> filter(title_general == 'Scientist') |>   table_basic_salary_stats()

}
