% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_career_progression}
\alias{plot_career_progression}
\title{plot career progression}
\usage{
plot_career_progression(
  .df,
  color_by = experience_highest_degree,
  color = c("#41AB5D", "goldenrod", "firebrick"),
  font_color = "#EEE8D5",
  bg_color = "black",
  font_size = 14,
  sizeref = 0.75,
  sizemin = 3,
  sizemax = 5
)
}
\arguments{
\item{.df}{salary data}
}
\description{
plot career progression
}
\examples{
library(rbiotechsalary)
salaries_sub <- salaries |>
  dplyr::filter(title_general \%in\% c('Scientist', 'Principal Scientist'),
         stringr::str_detect(location_country, 'United States') )
plot_career_progression(salaries_sub)

}
