% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_experience}
\alias{plot_experience}
\title{Plot salary against years of experience}
\usage{
plot_experience(.df, .plotly = TRUE)
}
\arguments{
\item{.df}{salary data}
}
\value{
ggplot
}
\description{
Plot salary against years of experience
}
\examples{
data(salaries)
salaries_sci <- salaries |>
  filter(title_general == 'Scientist', location_country == 'United States Of America')
plot_experience(salaries_sci)
}
