% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary}
\alias{plot_salary}
\title{plot salary histogram}
\usage{
plot_salary(
  .df,
  x = salary_total,
  fill = title_general,
  title = NULL,
  .plotly = TRUE
)
}
\arguments{
\item{.df}{salary data}
}
\description{
plot salary histogram
}
\examples{
library(r_biotech_salary)
data(salaries)
salaries_sub <- salaries |>
  filter(title_general == 'Scientist',
         stringr::str_detect(location_country, 'United States') )
plot_salary(salaries_sub, .plotly = FALSE)

}
