% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary_and_experience}
\alias{plot_salary_and_experience}
\title{Plot salary against years of experience}
\usage{
plot_salary_and_experience(.df)
}
\arguments{
\item{.df}{salary data}
}
\value{
ggplot
}
\description{
Plot salary against years of experience
}
\examples{
data(salaries)
salaries_sci <- salaries |> filter(title_general == 'Scientist')
plot_salary_and_experience(salaries_sci)
}
