% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary_title}
\alias{plot_salary_title}
\title{Make the title for salary histogram}
\usage{
plot_salary_title(.df, .return_data = FALSE, .gt = TRUE)
}
\arguments{
\item{.df}{salaries}
}
\description{
Make the title for salary histogram
}
\examples{
data(salaries)
salaries |>  plot_salary_title()

}
