% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rb_ui}
\alias{rb_ui}
\title{new ui}
\usage{
rb_ui()
}
\description{
new ui
}
\examples{
\dontrun{
  launch_app(ui = rb_ui2, server = function(input, output) {})
  launch_app(ui = rb_ui2, server = rb_server_2)
}
}
