% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_shinydashboard.R
\name{rb_ui2}
\alias{rb_ui2}
\title{new ui}
\usage{
rb_ui2()
}
\description{
new ui
}
\examples{
\dontrun{
  launch_app(ui = rb_ui2, server = function(input, output) {})
  launch_app(ui = rb_ui2, server = rb_server_2)
  
}
}
