% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_raw.R
\name{rt_table_raw}
\alias{rt_table_raw}
\title{reactable table showing raw data}
\usage{
rt_table_raw(.df, id = "table-raw", ...)
}
\arguments{
\item{id}{used for show/hide columns JS}
}
\description{
reactable table showing raw data
}
\examples{
salaries |> slice(1:20) |>  rt_table_raw()
}
