% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_basic_stats.R
\name{table_basic_salary_stats}
\alias{table_basic_salary_stats}
\title{small table salary stats}
\usage{
table_basic_salary_stats(.df)
}
\arguments{
\item{.df}{salaries}
}
\description{
small table salary stats
}
\examples{
data(salaries)
salaries |> filter(title_general == 'Scientist') |>   table_basic_salary_stats()

}
