% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_salary_stats.R
\name{table_salary_stats}
\alias{table_salary_stats}
\title{Create stats table for salary data}
\usage{
table_salary_stats(.df, .return_data = FALSE, .gt = TRUE)
}
\arguments{
\item{.df}{salaries}
}
\description{
Create stats table for salary data
}
\examples{
data(salaries)
salaries |> filter(title_general == 'Scientist') |>   table_salary_stats()

}
