% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_boxes.R
\name{value_boxes_stats_ui}
\alias{value_boxes_stats_ui}
\title{Value boxes ui}
\usage{
value_boxes_stats_ui(
  id,
  bg = "#1E2122",
  fg = "#41AB5DFF",
  showcase_layout = "left center",
  full_screen = TRUE,
  fill = TRUE,
  fillable = TRUE,
  height = NULL,
  max_height = "125px",
  ...
)
}
\description{
Value boxes ui
}
