% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_career_progression.R
\name{gt_career_progression}
\alias{gt_career_progression}
\title{table of summarized salary info}
\usage{
gt_career_progression(.df)
}
\value{
plot
}
\description{
not currently used
}
\examples{
gt_career_progression(salaries)
}
