% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_dark_mode.R
\name{gt_dark_mode}
\alias{gt_dark_mode}
\title{stylize to dark mode}
\usage{
gt_dark_mode(
  gt_object,
  .color_dark = "#FFFFFF00",
  .color_light = "#EEE8D5",
  ...
)
}
\arguments{
\item{gt_object}{gt table}

\item{...}{tab options}
}
\value{
gt table
}
\description{
stylize to dark mode
}
\examples{
  
salaries |> slice(1:20) |>  gt_table_raw() |>  gt_dark_mode('black','black')
salaries |> slice(1:20) |>  gt_table_raw() |>  
gt::tab_style(
style = gt::cell_fill(alpha = 0), locations = list(gt::cells_body()))  |> 
gt::opt_interactive()

}
