% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_shinydashboard.R, R/launch_app.R,
%   R/server.R
\name{rb_ui}
\alias{rb_ui}
\alias{launch_app}
\alias{rb_server}
\title{launch rbiotechsalary app}
\usage{
rb_ui()

launch_app(options = list(), ui = rb_ui, server = rb_server, ...)

rb_server(input, output, session)
}
\description{
launch rbiotechsalary app
}
\section{Functions}{
\itemize{
\item \code{rb_ui()}: UI components

\item \code{rb_server()}: updated server function

}}
\examples{
\dontrun{

  # launch with just ui
  launch_app(ui = rb_ui, server = function(input, output) {})

  # launch with server
  launch_app(ui = rb_ui, server = rb_server)
}
\dontrun{

 launch_app()
  
 # launch ui without server
 launch_app(server = function(input, output) {})
 
}
}
