% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{launch rbiotechsalary app}
\usage{
launch_app(...)
}
\arguments{
\item{server}{when not supplied, will use rb_server}
}
\description{
launch rbiotechsalary app
}
\examples{
\dontrun{
 
}
}
