% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grouped_options.R
\name{make_grouped_options}
\alias{make_grouped_options}
\title{Make grouped choices for shiny.fluent::dropDown}
\usage{
make_grouped_options(.df, .groups, .choices)
}
\arguments{
\item{.df}{salaries}

\item{.groups}{group column}

\item{.choices}{options column}
}
\description{
Make grouped choices for shiny.fluent::dropDown
}
\examples{
make_grouped_options(salaries, title_category, title_general)
}
