% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{make_hist_data}
\alias{make_hist_data}
\title{generates binned data and stats for histogram plot}
\usage{
make_hist_data(.df, x)
}
\description{
generates binned data and stats for histogram plot
}
\examples{
salaries |> make_hist_data(salary_total)
salaries |> filter(title_general == 'Scientist') |> make_hist_data(salary_total)

}
