% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_career_progression}
\alias{plot_career_progression}
\title{plot career progression}
\usage{
plot_career_progression(
  .df,
  x = date,
  fill = title_general,
  title = "",
  .type = "plotly"
)
}
\arguments{
\item{.df}{salary data}
}
\description{
plot career progression
}
\examples{
library(rbiotechsalary)
data(salaries)
salaries_sub <- salaries |>
  dplyr::filter(title_general \%in\% c('Scientist', 'Principal Scientist'),
         stringr::str_detect(location_country, 'United States') )
plot_salary(salaries_sub)

}
