% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary}
\alias{plot_salary}
\title{plot salary histogram}
\usage{
plot_salary(.df, x = salary_total, fill = title_general)
}
\arguments{
\item{.df}{salary data}
}
\description{
plot salary histogram
}
\examples{
library(r_biotech_salary)
data(salaries)
plot_salary(salaries)

}
