% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary_histogram}
\alias{plot_salary_histogram}
\title{plot salary histogram v2}
\usage{
plot_salary_histogram(
  .df,
  x,
  color = "#41AB5DFF",
  font_color = "#EEE8D5",
  hover_bg = "#161C21",
  ...
)
}
\description{
plot salary histogram v2
}
\examples{
plot_salary_histogram(salaries, x = salary_total, font_color = 'black', hover_bg = 'white')

salaries |>  filter(location_country == 'Canada', title_general == 'Associate Scientist') |> 
  plot_salary_histogram(x = salary_total, font_color = 'black', hover_bg = 'white')
  
salaries |>  slice(1) |> 
  plot_salary_histogram(x = salary_total, font_color = 'black', hover_bg = 'white')
}
