% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_salary.R
\name{plot_salary_histogram}
\alias{plot_salary_histogram}
\title{plot salary histogram v2}
\usage{
plot_salary_histogram(
  .plot_data,
  color = "#41AB5DFF",
  font_color = "#EEE8D5",
  hover_bg = "#161C21",
  ...
)
}
\description{
plot salary histogram v2
}
\examples{

# all
salaries |>  make_hist_data(salary_total) |>  plot_salary_histogram(font_color = 'black', hover_bg = 'white')

# scientist
salaries |>  filter(title_general == 'Scientist') |> make_hist_data(salary_total) |> plot_salary_histogram(font_color = 'black', hover_bg = 'white')
  
# 1
salaries |>  slice(1) |> make_hist_data(salary_total) |> 
  plot_salary_histogram(font_color = 'black', hover_bg = 'white')
}
