% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklines.R
\name{plot_sparkline}
\alias{plot_sparkline}
\title{sparkline for salary average showcase value box}
\usage{
plot_sparkline(.df, color = "white")
}
\description{
sparkline for salary average showcase value box
}
